Fix invisible obstacles and JSON stringify error
Basically, the issue was that the JSON request during simulation needs to read the locations of the obstacles in a very particular format, and calling remap to
reintroduce the obstacles into the randomization after they are hidden led the variable names of these obstacles to be different than what the JSON expected.
However, by setting the randomization obstacles back to the lastObstacles variable directly without remapping, the data can be read and plays nicely with the JSON request.