Change simulate request method from GET to POST
Previously, the simulate request was sent as a GET request, and the student code was sent as a query parameter. The request would fail when the student code was too long, because the URL was too long. Now, the simulate request is sent as a POST, so student code can be much longer.